Move view logic to MainView
